updates and refactorings
* updated pre-commit hooks
* migrate gitpod to ona
* updated and restructured nix stuff
* changed to conanfile.py
* updated python requirements
* updated build.sh and shebangs to be more portable
* enable warnings as errors by default
* updated ci pipelines
* updated tasks.json
* fixed presets.json
* updated/fixed README
* changed header file suffix to .h (Core Guideline recommendation)
* ci: fixed qt path on appveyor

+++ apps/qt/mainwindow.cpp
@@ -1,7 +1,7 @@
 #include "mainwindow.h"
 
-#include "coronan/corona-api_client.hpp"
-#include "country_chart_view.hpp"
+#include "coronan/corona-api_client.h"
+#include "country_chart_view.h"
 #include "ui_mainwindow.h"
 
 #include <QDebug>
@@ -62,9 +62,9 @@ void CoronanWidget::populate_date_boxes()
   auto const latest_date = latest_country_data.latest.date;
   // Unfortunatelly QDate(std::chrono::year_month_weekday_last date) can not be used when the compiler (libstdc++) does
   // not fully support C++ 20 even for Qt >= 6.4
-  auto const latest_qdate = QDate(static_cast<int>(latest_date.year()),       //
-                                  static_cast<unsigned>(latest_date.month()), //
-                                  static_cast<unsigned>(latest_date.day()));
+  auto const latest_qdate = QDate(static_cast<int>(latest_date.year()),            //
+                                  static_cast<int>(unsigned(latest_date.month())), //
+                                  static_cast<int>(unsigned(latest_date.day())));
 
   ui->startDate->setDate(default_start_date);
   ui->startDate->setMaximumDate(latest_qdate);
